Fix bug 1165434, use a global temp dir when no persists defined
When quitting mozgui after an app downloading for a bisection, when
persist is not defined by the user, the temp dir persists.
To fix this behaviour, we now create a temp_dir specific for the
mainwindow and pass it to the Bisector class.
Then on close, we explicly delete it.