Override default selection theme to keep cells verdict color
Cells Background color represents the verdict on the nightly by the user
of mozregui. But the selection default theme replace it with a blue background.

It hides a very useful information for the user.

The fix overrides this behavior and keeps verdict color on selected cells
but bolden its text.