Bug 1192485 - handle errors when running the tested application
This handle cases when we got some exception while:

 1. trying to install the application
 2. trying to start the application
 3. trying to stop the application.

For cases 1 and 2, this is an "uncoverable" error. So the raw
exception is printed, and we force skipping the build.

Case 3 can be recovered, so the only thing we do is print the raw
error, but we keep the verdict that the user entered and the
app_info dict extracted from the installed application.