Bug 1195622 - Error message is not clear when looking for inbounds using changesets that are in the same push
This handle tha case when a user give two changesets that are in
the same push. In that case, the pushlog url now only point to one
changeset and a message is printed at the end of the bisection to
inform the user.

Note that it also get rid of the "was_empty" attribute on
InboundBuildData. It is not required because this will be handled
in the bisection process. Also note that due to the implementation
of PushLogFinder, the returned list of push logs can never be empty
(at least one entry, else 404 HttpError is raised). This allow us
to simplify the error message when NO_DATA for inbounds.