Bug 1197568 - create a main.Application class
Replace the bisector.BisectRunner class with a main.Application class.

The purpose of this class is a bit more generic though it can still run
nightlies and inbounds bisection. Application is reponsible to create and
manage the required instances of BuildDownloadManager, TestRunner and Bisector.
It also create a temporary download dir if no persist dir is given.