Automatically bisect integration branch on merge
This change the workflow of mozregression (non GUI for now).

Basically, when we finish a nightly bisection, we switch to taskcluster
bisection to reduce the range on the same branch.

Once this is reduced to two push (that is one good, one bad), we parse
the log message and try to detect a merge. If we find a merge, we
continue the bisection on the branch where the merged commits comes from.