Bug 1217906 - allow to find build using build ids
This allow in --good-date, --bad-date, --launch to use a given buildid.

It works for nightlies only, as finding the buildid relies on the folder
name in archives.m.o.

A known bug is that files are stored using the build id, so it is
possible to have two files downloaded and kept in the persist dir
that are identical, example:

2015-11-03-03-02-48--mozilla-central--firefox-45.0a1.en-US.linux-x86_64.tar.bz2
2015-11-03--mozilla-central--firefox-45.0a1.en-US.linux-x86_64.tar.bz2