fix fennec get_nightly_repo
For fennec, get_nightly repo was used to return the part of the regex to
find the build url.

This fixes it, now get_nightly_repo_regex does that job. Note that it
should not have a functional impact, but this is cleaner and will be
required when we will implement the 'new workflow'.