Bug 1229976 - use json-pushes to determine arange
Uses json-pushes API to determine the date range. Before we were trying
to find such a range by trying to map a date to a changeset - but this
is error prone (in case we have no changeset for date limit) and we were
doing three requests.

Now only one request is sufficient - and we let json-pushes handle the
corner cases described above.

Also, I added some logging statements. And now pushes
from *pushlog_within_changes* are sorted by pushid, instead of
date (note that it should not change anything, but this is more logical).