Bug 1231745 - fix race condition when removing dl dir
Ensure that we don't have a race condition when the download_dir is
removed by first waiting for download threads to be finished.

This is not required for the GUI because we wait for threads completion
then we remove the download dir in a synchronized way in the aboutToQuit
application's slots. Also we don't need that in the command line when
we don't remove the download dir because python automatically wait
for those threads before exiting.