Bug 1245333 - [gui] wrong build for retry
We calculate the middle of the build range to know which build should be
run.
In case a retry is done after a skip, this lead to an error - so for the
retry action we should not calculate the middle again, just reuse the
previous one.

I took the opportunity to clean the mid/self.mid redundancy in the code.