Bug 1245332 - stopping firefox may hang on windows
On windows, if a firefox build is stopped after it had restarted (from
Menu (≡) -> [?] -> Restart with Add-ons Disabled for example) it will
hang for a couple of minutes.

The blocking comes from mozprocess, but it is quite hard to fix properly,
and it is a corner case - so here is a workaround patch.

It run the stop method in a thread, and wait 0.7 seconds for its
completion. If the thread is not terminated after that, we simply forget
it, so it will be finished later (abruptly in case the program ends
before the completion).