Don't expose fallback routes in warning-level logs.
Previously when mozregression failed to find any builds for a revision (because
of build failures, or in this case builds not yet being complete) after doing
route fallback it would print a log message about failing to find build info at
the final route it tried (pgo).
This caused actual confusion for a user who thought GUI had ignored the choice
of shippable. Log the original requested route instead.