Bug 1404520 - Raise a helpful error when missing required build info
The NightlyInfoFetcher._fetch_build_info_from_url() method constructs
a dict with at least one key, "build_url", that is passed back
to the caller. However it was possible under specific circumstances to
build a dict where the "build_url" key was missing, causing later parts
of the program to fail.

This code adds a check for the "build_url" key and raises a helpful
error if the key is missing.