Bug 1589881 - Don't fail if a pushlog URL 404's (#558)
Currently if a push doesn't exist in a repo we look for it in, the 404 error
results in a `MozRegressionError` exception, which is fatal. That means we're
failing out if a push doesn't exist in any one of the repos we search, even if
it does exist in another (for example, an autoland push that's been merged to
central, but not yet back to inbound). So, change that exception to an
`EmptyPushlogError`, which isn't fatal.

I'd expect this to be a pretty common problem now that merges to
mozilla-inbound have been slowed down and commit access to it is being sharply
limited.