Bug 1401649 - Add a new 'firefox-l10n' app that bisects Fx l10n nightlies
firefox-l10n nightlies are awkward to bisect with mozregression since they
don't have a build info .txt file in their build directory, instead one has to
fetch it from the build dir of the regular en-US build.  Thus we need to
introduce a new fetch_config property, `has_build_info`, to advertise that fact,
and when it's false fetch the info separately.