Return full URLs from `url_links()` and use them when searching for builds
The way url_links was returning only the leaf node of the path rather than full
URLs was making it awkward to implement the firefox-l10n app, which needs to use
the function for two different URLs (one for the folder with the l10n builds,
and the other for the regular m-c nightly folder for the build info .txt file).

I also found it unpleasant on principle, especially given the result is then
immediately simply concatenated with the page URL to get an (hopefully) absolute
URL anyway, without proper URL parsing (pretty much guaranteed to work given the
known page structure to be fair, but still felt icky).  The new way is still not
perfect since there's no accounting for any hypothetical baseURI, but that seems
a distant concern.