launchers: Disable TB update by policy (bug 1783694)
Although the `app.update.auto:false` and `app.update.enabled:false`
preferences are set, these have no effect on recent versions of
Thunderbird due to their removal in Bug 1420514 (ported to TB in Bug
1479289).  The replacement policy mechanism of setting
`policies.DisableAppUpdate:true` in `distribution/policies.json`
was added to FirefoxLauncher in 68ece936, but was not added to
ThunderbirdLauncher.  This commit applies the same change so that app
updates are disabled for Thunderbird.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>