Bug 1857026 - Add integration branch bisection for Fenix/Focus
Set the name of the build type to "shippable" since the builds are all PGO. Note
that the task names for Fenix call themselved 'nightly-simulation' but the intent
is the same as 'shippable'.

This also adds support for Focus, but that is less valuable since we don't
schedule compatible builds very often.